+++ mpv.net/Native/Native.cs
@@ -27,6 +27,9 @@ namespace mpvnet
         [DllImport("user32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
         public static extern IntPtr PostMessage(IntPtr hWnd, int Msg, IntPtr wParam, IntPtr lParam);
 
+        [DllImport("user32.dll", CharSet = CharSet.Unicode)]
+        public static extern int RegisterWindowMessage(string id);
+
         [DllImport("user32.dll")]
         public static extern bool AllowSetForegroundWindow(int dwProcessId);