+++ Changelog.md
@@ -1,5 +1,6 @@
 ###
 
+- added taskbar-progress implementation
 - added new setting to start with maximized window
 - long file names work now even if not enabled by the OS
 - fixed history being written even when history file wasn't created prior